Major work.
- Splitted CanvasParser
- Cleaned up datamodel code
- Fixed most of the warnings.
- Added conditonal style classes.

+++ NuakeUI/Font.cpp
@@ -29,7 +29,7 @@ namespace NuakeUI
 		msdfgen::FontMetrics metrics;
 		msdfgen::getFontMetrics(metrics, mFontHandle);
 
-		LineHeight = metrics.lineHeight;
+		LineHeight = (float)metrics.lineHeight;
 		return true;
 	}
 }
